Port the "pretty" module.
Includes an abridged port of the Haskell `Text.PrettyPrint.HughesPJ`
module, with just the functionality needed for pretty-printing out
C ASTs.

A more faithful port would be easy to write down, but hard to finish
since it entails the kind of lazy manipulation of recursive structs
that is common in Haskell code, but leads to abysmal Rust runtime
behavior when translated naively.